backend/requirements.txt
# Django core
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-filter>=24.0

# MongoDB (remplace PostgreSQL)
django-mongodb-backend>=5.1
pymongo[srv]>=4.7
dnspython>=2.6        # requis pour les URI mongodb+srv://

# Fichiers statiques (Render, sans Nginx)
whitenoise>=6.7

# Emails
resend>=2.0

# Async tasks
celery>=5.4
redis>=5.0
django-celery-beat>=2.6

# API documentation
drf-spectacular>=0.27

# Production server
gunicorn>=22.0

# Utilities
python-dotenv>=1.0
